[FEATURE modernized-built-in-components] Extract shared code
Refactor to extra shared code into internal component to pave way
for implementing <LinkTo> with the new infrastructure.

Part of #19270
