+++ packages/@ember/-internals/glimmer/index.ts
@@ -366,9 +366,10 @@ export { templateFactory as template, templateCacheCounters } from '@glimmer/opc
 export { default as RootTemplate } from './lib/templates/root';
 export { default as Checkbox } from './lib/components/checkbox';
 export { default as TextField } from './lib/components/text-field';
-export { default as TextArea } from './lib/components/textarea';
+export { default as TextArea } from './lib/components/-textarea';
 export { default as LinkComponent } from './lib/components/link-to';
-export { default as Input, TextareaComponent as Textarea } from './lib/components/input';
+export { default as Input } from './lib/components/input';
+export { default as Textarea } from './lib/components/textarea';
 export { default as Component } from './lib/component';
 export { default as Helper, helper } from './lib/helper';
 export { SafeString, escapeExpression, htmlSafe, isHTMLSafe } from './lib/utils/string';
