+++ packages/@ember/-internals/glimmer/lib/setup-registry.ts
@@ -2,11 +2,12 @@ import { privatize as P, Registry } from '@ember/-internals/container';
 import { ENV } from '@ember/-internals/environment';
 import { EMBER_MODERNIZED_BUILT_IN_COMPONENTS } from '@ember/canary-features';
 import Component from './component';
+import LegacyTextArea from './components/-textarea';
 import Checkbox from './components/checkbox';
-import InputComponent, { TextareaComponent } from './components/input';
-import LinkToComponent from './components/link-to';
+import Input from './components/input';
+import LinkTo from './components/link-to';
 import TextField from './components/text-field';
-import TextArea from './components/textarea';
+import Textarea from './components/textarea';
 import { clientBuilder, rehydrationBuilder, serializeBuilder } from './dom';
 import loc from './helpers/loc';
 import { Renderer } from './renderer';
@@ -57,14 +58,14 @@ export function setupEngineRegistry(registry: Registry): void {
 
   registry.register('component:-text-field', TextField);
   registry.register('component:-checkbox', Checkbox);
-  registry.register('component:input', InputComponent);
-  registry.register('component:link-to', LinkToComponent);
+  registry.register('component:input', Input);
+  registry.register('component:link-to', LinkTo);
 
   if (EMBER_MODERNIZED_BUILT_IN_COMPONENTS) {
-    registry.register('component:-textarea', TextArea);
-    registry.register('component:textarea', TextareaComponent);
+    registry.register('component:-textarea', LegacyTextArea);
+    registry.register('component:textarea', Textarea);
   } else {
-    registry.register('component:textarea', TextArea);
+    registry.register('component:textarea', LegacyTextArea);
   }
 
   if (!ENV._TEMPLATE_ONLY_GLIMMER_COMPONENTS) {
