+++ packages/@ember/-internals/glimmer/lib/templates/textarea.hbs
@@ -40,7 +40,7 @@
     {{on "paste" this.valueDidChange}}
     {{on "cut" this.valueDidChange}}
 
-    {{this.DeprecatedEventHandlersModifier this}}
+    {{this.handleDeprecatedEvents this}}
   />
 {{~else~}}
   {{~#let (component '-textarea')  as |Textarea|~}}
